Created a new method in subject-service instead of keeping it in the component
