+++ src/services/subject-service.ts
@@ -1,5 +1,5 @@
 import { axiosClient } from '~/plugins/axiosClient'
-import { AxiosResponse } from 'axios'
+import { AxiosResponse, AxiosRequestConfig } from 'axios'
 
 import { URLs } from '~/constants/request'
 import { ItemsWithCount, SubjectInterface, SubjectNameInterface } from '~/types'
@@ -18,5 +18,24 @@ export const subjectService = {
   ): Promise<AxiosResponse<SubjectNameInterface[]>> => {
     const category = createUrlPath(URLs.categories.get, categoryId)
     return axiosClient.get(`${category}${URLs.subjects.getNames}`)
+  },
+
+  sendSubjectRequest: (): Promise<AxiosResponse> => {
+    const config: AxiosRequestConfig = {
+      url: 'https://api.example.com/subjects',
+      method: 'get',
+      headers: {
+        Authorization: 'Bearer your-token-here',
+        'Content-Type': 'application/json'
+      }
+    }
+
+    return Promise.resolve({
+      data: { message: 'Success' },
+      status: 200,
+      statusText: 'OK',
+      headers: {},
+      config
+    } as AxiosResponse)
   }
 }
